+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/TeamInvestmentCompletedEventDTO.kt
@@ -6,7 +6,7 @@ data class TeamInvestmentCompletedEventDTO(
     val teamId: UUID,
     val teamName: String?,
     val curInvRound: Int,
-    val totalMoney: Long,
-    val valuationMoney: Long,
+    val totalMoney: Int,
+    val valuationMoney: Int,
     val profitNum: Double
 )
